Apply final review cleanups for instance handling changes
Agent-Logs-Url: https://github.com/ArduPilot/apm_planner/sessions/95dc171e-77d8-45f9-8de1-d7ac20000f1c

+++ src/ui/Loghandling/LogdataStorage.cpp
@@ -214,7 +214,7 @@ bool LogdataStorage::addDataRow(const QString &typeName, const QList<QPair<QStri
         newRow.m_values.push_back(values[i].second);
     }
 
-    if((tempType.m_indexFieldIndex >= 0) && (tempType.m_indexFieldIndex < newRow.m_values.size()))
+    if((tempType.m_indexFieldIndex >= 0) && (tempType.m_indexFieldIndex < static_cast<int>(newRow.m_values.size())))
     {
         bool ok = false;
         const int messageIndex = newRow.m_values.at(tempType.m_indexFieldIndex).toInt(&ok);
